requirements.txt
flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
Pillow>=10.0.1
waitress==3.0.2